Merge branch 'develop' into userbug12

Conflicts:
	Goluk/libs/armeabi/libgolukmobile.so
	Goluk/src/cn/com/mobnote/golukmobile/MainActivity.java

+++ Goluk/src/cn/com/mobnote/golukmobile/VideoShareActivity.java
@@ -40,6 +40,7 @@ import android.widget.Spinner;
 import android.widget.Toast;
 import cn.com.mobnote.application.GlobalWindow;
 import cn.com.mobnote.application.GolukApplication;
+import cn.com.mobnote.golukmobile.carrecorder.util.ImageManager;
 import cn.com.mobnote.logic.GolukModule;
 import cn.com.mobnote.module.page.IPageNotifyFn;
 import cn.com.mobnote.umeng.widget.CustomShareBoard;
@@ -105,7 +106,7 @@ public class VideoShareActivity extends Activity implements OnClickListener {
 	public static final String PATH_FS1 = "/tiros-com-cn-ext";
 
 	final String fileFolder = Environment.getExternalStorageDirectory().getPath() + PATH_FS1 + "/Cache";
-	final String thumbFile = fileFolder + "/thumb11.jpg";
+	private String thumbFile = fileFolder + "/thumb11.jpg";
 
 	/** 上传视频更新进度 */
 	private final int MSG_H_UPLOAD_PROGRESS = 2;
@@ -165,6 +166,8 @@ public class VideoShareActivity extends Activity implements OnClickListener {
 	SharePlatformUtil sharePlatform;
 	/** 2/3 紧急/精彩 */
 	private int mVideoType = 0;
+	
+	private String videoName ;
 
 	public Handler mmmHandler = new Handler() {
 		public void handleMessage(android.os.Message msg) {
@@ -295,12 +298,12 @@ public class VideoShareActivity extends Activity implements OnClickListener {
 		Intent intent = getIntent();
 		mVideoPath = intent.getStringExtra("cn.com.mobnote.golukmobile.videopath");
 		mVideoType = intent.getIntExtra("type", 0);
-
+		videoName = intent.getStringExtra("videoName");
 		// 获得GolukApplication对象
 		mApp = (GolukApplication) getApplication();
 		mApp.setContext(this, "VideoShare");
 		sharePlatform = new SharePlatformUtil(this);
-		
+		sharePlatform.configPlatforms();// 设置分享平台的参数
 		// 获取第一帧缩略图
 		createThumb();
 		// 初始化
@@ -313,42 +316,54 @@ public class VideoShareActivity extends Activity implements OnClickListener {
 
 	private void createThumb() {
 		long startTime = System.currentTimeMillis();
-		mShortBitmap = ThumbnailUtils.createVideoThumbnail(mVideoPath, Thumbnails.MINI_KIND);
-		if (mShortBitmap != null) {
-			int width = mShortBitmap.getWidth();
-			int height = mShortBitmap.getHeight();
+		//mShortBitmap = ThumbnailUtils.createVideoThumbnail(mVideoPath, Thumbnails.MINI_KIND);
+			String filePath = GolukApplication.getInstance().getCarrecorderCachePath() + File.separator + "image";
+			thumbFile = filePath + File.separator + videoName;
+			mShortBitmap = ImageManager.getBitmapFromCache(thumbFile, 194, 109);
+			
+			File image = new File(thumbFile);
+			if(image.exists()){
+				return ;
+			}else{
+				thumbFile = fileFolder + "/thumb11.jpg";
+				mShortBitmap = ThumbnailUtils.createVideoThumbnail(mVideoPath, Thumbnails.MINI_KIND);
+				if (mShortBitmap != null) {
+					int width = mShortBitmap.getWidth();
+					int height = mShortBitmap.getHeight();
 
-			Log.e("", "VideoShareActivity createThumb: width:" + width + "	height:" + height);
-		} else {
-			Log.e("", "VideoShareActivity createThumb: NULL:");
-		}
+					Log.e("", "VideoShareActivity createThumb: width:" + width + "	height:" + height);
+				} else {
+					Log.e("", "VideoShareActivity createThumb: NULL:");
+				}
 
-		try {
+				try {
 
-			File file = new File(fileFolder);
-			file.mkdirs();
-			file = new File(thumbFile);
-			if (file.exists()) {
-				file.delete();
+					File file = new File(fileFolder);
+					file.mkdirs();
+					file = new File(thumbFile);
+					if (file.exists()) {
+						file.delete();
+					}
+
+					file.createNewFile();
+
+					FileOutputStream fos = new FileOutputStream(file);
+					mShortBitmap.compress(Bitmap.CompressFormat.JPEG, 100, fos);
+
+					String fsFile = FileUtils.javaToLibPath(thumbFile);
+
+					Log.e("", "VideoShareActivity createThumb: time: " + fsFile);
+
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 			}
-
-			file.createNewFile();
-
-			FileOutputStream fos = new FileOutputStream(file);
-			mShortBitmap.compress(Bitmap.CompressFormat.JPEG, 100, fos);
-
-			String fsFile = FileUtils.javaToLibPath(thumbFile);
-
-			Log.e("", "VideoShareActivity createThumb: time: " + fsFile);
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
 
 		long dur = System.currentTimeMillis() - startTime;
 
 		Log.e("", "VideoShareActivity createThumb: time:" + dur);
 	}
+	
 
 	private void dimissErrorDialog() {
 		if (null != mErrorDialog) {
@@ -668,7 +683,8 @@ public class VideoShareActivity extends Activity implements OnClickListener {
 
 	private void click_share() {
 		if (!this.mIsUploadSucess) {
-			Toast.makeText(VideoShareActivity.this, "上传视频成功后才可以分享", Toast.LENGTH_SHORT).show();
+			//Toast.makeText(VideoShareActivity.this, "上传视频成功后才可以分享", Toast.LENGTH_SHORT).show();
+			Toast.makeText(VideoShareActivity.this, "正在上传视频,请稍等", Toast.LENGTH_SHORT).show();
 			return;
 		}
 
@@ -698,13 +714,13 @@ public class VideoShareActivity extends Activity implements OnClickListener {
 			Toast.makeText(VideoShareActivity.this, "第三方分享失败", Toast.LENGTH_SHORT).show();
 			return;
 		}
-		Toast.makeText(VideoShareActivity.this, "开始第三方分享:" + channel, Toast.LENGTH_SHORT).show();
+		//Toast.makeText(VideoShareActivity.this, "开始第三方分享:" + channel, Toast.LENGTH_SHORT).show();
 
 		final String json = createShareSucesNotifyJson(mVideoVid, channel);
 		boolean b = mApp.mGoluk.GolukLogicCommRequest(GolukModule.Goluk_Module_HttpPage,
 				IPageNotifyFn.PageType_ShareNotify, json);
 
-		Toast.makeText(VideoShareActivity.this, "调用Logic分享结果:" + json, Toast.LENGTH_SHORT).show();
+		//Toast.makeText(VideoShareActivity.this, "调用Logic分享结果:" + json, Toast.LENGTH_SHORT).show();
 
 		if (!b) {
 			Toast.makeText(VideoShareActivity.this, "调用Logic分享结果失败:" + channel, Toast.LENGTH_SHORT).show();
